# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Koupreng E-Invitation API
  description: |
    REST API for the Koupreng E-Invitation platform, including authentication,
    users, invitations, templates, guests, RSVP, check-in, media, payments,
    organizations, subscriptions, notifications, reporting, administration,
    and related services. The current contract includes both legacy /api routes
    and versioned /api/v1 routes; the version below does not imply a global base path.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
...
paths:
...
  /api/v1/admin/templates/{templateId}:
    get:
      tags:
      - Administration
      summary: Get Template
      operationId: getTemplate
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminTemplateResponse"
      security:
      - bearerAuth: []
    put:
      tags:
      - Administration
      summary: Update Template
      operationId: updateTemplate
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminTemplateRequest"
            examples:
              demoTemplate:
                summary: Create a fictional wedding template
                value:
                  name: Koupreng Demo Wedding
                  code: koupreng-demo-wedding
                  category: TRADITIONAL
                  description: A fictional bilingual Khmer wedding template.
                  thumbnailUrl: https://example.com/images/demo-template-thumb.webp
                  previewUrl: https://example.com/images/demo-template-preview.webp
                  price: 0.0
                  currency: USD
                  premium: false
                  status: ACTIVE
                  sortOrder: 100
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminTemplateResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Administration
      summary: Delete Template
      operationId: deleteTemplate
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
      security:
      - bearerAuth: []
...
  /api/v1/template-payments/static/create:
    post:
      tags:
      - Payments
      summary: Create a static ABA payment order
      operationId: createStaticPaymentOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplatePaymentRequest"
            examples:
              pendingStaticOrder:
                summary: Create a pending static ABA order; this does not confirm
                  payment
                value:
                  templateId: 1
                  templateName: Koupreng Demo Wedding
                  packageName: Single Template
                  amount: 0.01
                  currency: USD
                  buyerName: Koupreng Demo User
                  buyerEmail: demo@koupreng.local
                  buyerPhone: "012345678"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCreateTemplatePaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/template-payments/payway/create:
    post:
      tags:
      - Payments
      summary: Create a PayWay template checkout
      operationId: createPaywayCheckout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplatePaymentRequest"
            examples:
              pendingStaticOrder:
                summary: Create a pending static ABA order; this does not confirm
                  payment
                value:
                  templateId: 1
                  templateName: Koupreng Demo Wedding
                  packageName: Single Template
                  amount: 0.01
                  currency: USD
                  buyerName: Koupreng Demo User
                  buyerEmail: demo@koupreng.local
                  buyerPhone: "012345678"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCreateTemplatePaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/template-payments/{orderCode}/claim:
    post:
      tags:
      - Payments
      summary: Request payment review
      description: A buyer claim is not payment proof. Only trusted confirmation can
        unlock the template.
      operationId: claimPayment
      parameters:
      - name: orderCode
        in: path
        required: true
        schema:
          type: string
        example: <pending-order-code>
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimTemplatePaymentRequest"
            examples:
              paymentReference:
                summary: Claim an owned pending payment
                value:
                  reference: ABA-TRANSACTION-REFERENCE
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentConfirmResponse"
      security:
      - bearerAuth: []
  /api/v1/template-payments/orders/{orderCode}/claim:
    post:
      tags:
      - Payments
      summary: Request payment review
      description: A buyer claim is not payment proof. Only trusted confirmation can
        unlock the template.
      operationId: claimPayment_1
      parameters:
      - name: orderCode
        in: path
        required: true
        schema:
          type: string
        example: <pending-order-code>
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimTemplatePaymentRequest"
            examples:
              paymentReference:
                summary: Claim an owned pending payment
                value:
                  reference: ABA-TRANSACTION-REFERENCE
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentConfirmResponse"
      security:
      - bearerAuth: []
  /api/v1/template-payments/create:
    post:
      tags:
      - Payments
      summary: Create a static ABA template payment
      operationId: createStaticPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplatePaymentRequest"
            examples:
              pendingStaticOrder:
                summary: Create a pending static ABA order; this does not confirm
                  payment
                value:
                  templateId: 1
                  templateName: Koupreng Demo Wedding
                  packageName: Single Template
                  amount: 0.01
                  currency: USD
                  buyerName: Koupreng Demo User
                  buyerEmail: demo@koupreng.local
                  buyerPhone: "012345678"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCreateTemplatePaymentResponse"
      security:
      - bearerAuth: []
...
  /api/v1/payway/callback:
    post:
      tags:
      - Payments
      summary: Receive a PayWay payment callback
      description: Provider-facing callback. The backend validates the PayWay signature
        and transaction before changing access.
      operationId: paywayCallback
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePayWayCallbackResponse"
      security: []
...
  /api/v1/internal/subscription-payments/telegram-detect:
    post:
      tags:
      - Subscriptions
      summary: Reconcile trusted Telegram evidence for a subscription payment
      operationId: detectInternalTelegramPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramDetectSubscriptionPaymentRequest"
            examples:
              trustedTelegramEvidence:
                summary: Reconcile a merchant Telegram notification
                value:
                  rawMessage: "$199.00 paid by KOEURNG VIREAK (*247). Trx. ID: 178002499089682,\
                    \ APV: 383331."
                  detectedBy: telegram-admin-detect:123456789
                  telegramChatId: "-1001234567890"
                  telegramMessageId: "42"
                  telegramSenderId: "987654321"
                  telegramSenderUsername: PayWayByABA_bot
                  detectedAmount: 199.0
                  detectedCurrency: USD
                  payerName: KOEURNG VIREAK
                  payerAccountLast3: "247"
                  paywayTransactionId: "178002499089682"
                  paywayApprovalCode: "383331"
                  remark: null
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSubscriptionPaymentDetectionResponse"
      security:
      - adminPaymentSecret: []
...
  /api/v1/admin/templates:
    get:
      tags:
      - Administration
      summary: List Templates
      operationId: listTemplates
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListAdminTemplateResponse"
      security:
      - bearerAuth: []
    post:
      tags:
      - Administration
      summary: Create Template
      operationId: createTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminTemplateRequest"
            examples:
              demoTemplate:
                summary: Create a fictional wedding template
                value:
                  name: Koupreng Demo Wedding
                  code: koupreng-demo-wedding
                  category: TRADITIONAL
                  description: A fictional bilingual Khmer wedding template.
                  thumbnailUrl: https://example.com/images/demo-template-thumb.webp
                  previewUrl: https://example.com/images/demo-template-preview.webp
                  price: 0.0
                  currency: USD
                  premium: false
                  status: ACTIVE
                  sortOrder: 100
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminTemplateResponse"
      security:
      - bearerAuth: []
  /api/v1/admin/template-payments/telegram-detect:
    post:
      tags:
      - Payments
      summary: Review Telegram-detected payment data
      description: Requires a bearer JWT for a user with the ADMIN role.
      operationId: detectTelegramPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramDetectPaymentRequest"
            examples:
              adminReview:
                summary: Safe placeholder requiring administrator authorization
                value:
                  rawMessage: <verified-provider-message>
                  detectedBy: admin-demo
                  telegramChatId: <telegram-chat-id>
                  telegramMessageId: <telegram-message-id>
                  telegramSenderUsername: <telegram-username>
                  telegramSenderId: <telegram-sender-id>
                  detectedOrderCode: <pending-order-code>
                  detectedAmount: 0.01
                  detectedCurrency: USD
                  paywayTransactionId: null
                  paywayApprovalCode: null
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentConfirmResponse"
      security:
      - bearerAuth: []
  /api/v1/admin/template-payments/confirm:
    post:
      tags:
      - Payments
      summary: Confirm a template payment manually
      description: Requires a bearer JWT for a user with the ADMIN role.
      operationId: confirmManualPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmTemplatePaymentRequest"
            examples:
              adminReview:
                summary: Requires a real pending order and administrator authorization
                value:
                  orderCode: <pending-order-code>
                  amount: 0.01
                  confirmedBy: admin-demo
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentConfirmResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/payments/confirm:
    post:
      tags:
      - Payments
      summary: Confirm and fulfill a payment manually
      description: Confirms a server-priced template or subscription order. Requires
        ADMIN authority.
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentRequest"
            examples:
              subscriptionReview:
                summary: Confirm a real server-priced order after administrator review
                value:
                  orderCode: <pending-order-code>
                  amount: 19.0
                  confirmedBy: admin-demo
                  itemType: SUBSCRIPTION
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentConfirmResponse"
      security:
      - bearerAuth: []
...
  /api/v1/templates/{templateId}:
    get:
      tags:
      - Templates
      summary: Get Template
      operationId: getTemplate_1
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePublicTemplateResponse"
...
  /api/v1/template-payments/{orderCode}:
    get:
      tags:
      - Payments
      summary: Get an owned template payment order
      operationId: getOrder
      parameters:
      - name: orderCode
        in: path
        required: true
        schema:
          type: string
        example: <pending-order-code>
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTemplatePaymentStatusResponse"
      security:
      - bearerAuth: []
...
  /api/v1/notifications/summary:
    get:
      tags:
      - Notifications
      summary: Summarize notification activity
      operationId: summary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseNotificationSummaryResponse"
      security:
      - bearerAuth: []
  /api/v1/me/templates/{templateId}/access:
    get:
      tags:
      - Payments
      summary: Check the current user's template access
      operationId: templateAccess
      parameters:
      - name: templateId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTemplateAccessCheckResponse"
      security:
      - bearerAuth: []
  /api/v1/me/templates/paid:
    get:
      tags:
      - Payments
      summary: List templates paid for by the current user
      operationId: paidTemplates
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListUserTemplateAccessResponse"
      security:
      - bearerAuth: []
...
  /api/v1/me/subscription:
    get:
      tags:
      - Subscriptions
      summary: Retrieve the current subscription
      operationId: current
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSubscriptionResponse"
      security:
      - bearerAuth: []
  /api/v1/me/subscriptions/current:
    get:
      tags:
      - Subscriptions
      summary: Retrieve the current subscription
      operationId: current_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSubscriptionResponse"
      security:
      - bearerAuth: []
  /api/v1/me/payments:
    get:
      tags:
      - Payment History
      summary: List Mine
      operationId: listMine_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPaymentHistoryResponse"
      security:
      - bearerAuth: []
  /api/v1/me/payments/{orderCode}:
    get:
      tags:
      - Payment History
      summary: Retrieve a payment history record
      operationId: get_3
      parameters:
      - name: orderCode
        in: path
        required: true
        schema:
          type: string
        example: <pending-order-code>
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentHistoryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/seating/summary:
    get:
      tags:
      - Seating
      summary: Summarize seating activity
      operationId: summary_1
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSeatingSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/rsvps/summary:
    get:
      tags:
      - RSVP
      summary: Summarize rsvp activity
      operationId: summary_2
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseRsvpSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/invitations/{invitationId}/check-in/summary:
    get:
      tags:
      - Check-In
      summary: Summarize check in activity
      operationId: summary_4
      parameters:
      - name: invitationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCheckInSummaryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/template-payments:
    get:
      tags:
      - Payments
      summary: List template payment orders
      description: Requires a bearer JWT for a user with the ADMIN role.
      operationId: adminOrders
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListTemplatePaymentStatusResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/reports/payments:
    get:
      tags:
      - Administration
      summary: Payments Report
      operationId: paymentsReport
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminReportResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/payments:
    get:
      tags:
      - Administration
      summary: List All Payments
      operationId: listAllPayments
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListPaymentHistoryResponse"
      security:
      - bearerAuth: []
  /api/v1/admin/payments/{orderCode}:
    get:
      tags:
      - Administration
      summary: Get Payment
      operationId: getPayment
      parameters:
      - name: orderCode
        in: path
        required: true
        schema:
          type: string
        example: <pending-order-code>
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePaymentHistoryResponse"
      security:
      - bearerAuth: []
...
  /api/v1/admin/analytics/templates:
    get:
      tags:
      - Administration
      summary: Analytics Templates
      operationId: analyticsTemplates
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseAdminReportResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    AdminTemplateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
        category:
          type: string
          enum:
          - MODERN
          - TRADITIONAL
          - MINIMALIST
          - FLORAL
          - LUXURY
          - OTHER
        description:
          type: string
        thumbnailUrl:
          type: string
        previewUrl:
          type: string
        price:
          type: number
        currency:
          type: string
        premium:
          type: boolean
        status:
          type: string
        sortOrder:
          type: integer
          format: int32
        primaryColor:
          type: string
        secondaryColor:
          type: string
        backgroundColor:
          type: string
      required:
      - name
...
    ApiResponseAdminTemplateResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/AdminTemplateResponse"
        timestamp:
          type: string
          format: date-time
...
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste the accessToken returned by POST /api/auth/login. Scalar
        adds the Bearer prefix automatically; do not type it twice.
      scheme: bearer
      bearerFormat: JWT
    adminPaymentSecret:
      type: apiKey
      description: Internal Telegram payment reconciliation secret.
      name: X-ADMIN-PAYMENT-SECRET
      in: header